% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_dependent.R
\name{contrast_dependent}
\alias{contrast_dependent}
\title{Contrast analyses for dependent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
contrast_dependent(nGroup, lambda, dat, testvalue = 0)
}
\arguments{
\item{nGroup}{Number of dependent / within-subject groups}

\item{lambda}{A matrix of contrast weights with contrasts in rows and groups in
columns}

\item{dat}{A matrix or dataframe with nGroup columns; each row contains values for
one respondent;
each column contains values of the dependent variable in the respective within-
subject group}

\item{testvalue}{Value under the null hypothesis; if not specified, it is fixed to 0.}
}
\value{
a dataframe with following entries for each of the contrasts:
\describe{
  \item{\code{SumsofSquares}}{Sums of Squares}
  \item{\code{F}}{F-values}
  \item{\code{estimate}}{Contrast estimates}
  \item{\code{t}}{t-values}
  \item{\code{p}}{Two-tailed p-values}
  \item{\code{g}}{Effect size g, a standardized distance measure}
}
}
\description{
This function allows to perform contrast analyses for dependent samples.
Please note that sample sizes must be equal between within-subject groups,
so no missings are allowed
}
\examples{
# Load presidents dataset
dat <- data.frame(Qtr1=presidents[seq(1, length(presidents), 4)],
               Qtr2=presidents[seq(2, length(presidents), 4)],
               Qtr3=presidents[seq(3, length(presidents), 4)],
               Qtr4=presidents[seq(4, length(presidents), 4)])
               dat <- na.omit(dat)

nGroup <- ncol(dat)
# Define lambda weights
lambda <- matrix(c(
    1, 0, 0, -1, # H1: decrease in approval ratings with stagnation over warmer months
    3, 1, -1, -3), # H2: linear decrease in  approval ratings
    ncol = nGroup,
    byrow=TRUE)

# Perform contrast analysis
contrast_dependent(nGroup, lambda,dat)
# t > 2 indicates that Contrast 1 and Contrast 2 fit the data well
# Both tests are significant with p < .05. This suggests that presidental approval
# ratings decrease over the four calendar quarters, but it cannot be identified
# whether or not this decrease stagnates over warmer months.

}
