% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_independent.R
\name{contrast_independent}
\alias{contrast_independent}
\title{Contrast analyses for independent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
contrast_independent(nGroup, lambda, dat)
}
\arguments{
\item{nGroup}{Number of independent / between-subject groups}

\item{lambda}{A matrix of contrast weights with contrasts in rows and groups in
columns}

\item{dat}{A matrix or dataframe with two columns; each row contains values for
one respondent;
the first column contains the group indicator, the second column contains the
dependent variable}
}
\value{
A dataframe with following entries for each of the contrasts:
\describe{
  \item{\code{SumsofSquares}}{Sums of Squares}
  \item{\code{F}}{F-values}
  \item{\code{estimate}}{Contrast estimates}
  \item{\code{t}}{t-values}
  \item{\code{p}}{Two-tailed p-values}
  \item{\code{rEffectSize}}{Correlation between the dependent variable and the
   contrast weights}
  \item{\code{rAlerting}}{Correlation between group means and contrast weights}
  \item{\code{r2Aalerting}}{Squared \code{rAlerting}; can be interpreted similar to
   a determination coefficient as a measured of explained variance by the contrast
  \code{(SScontrast/SSbetween)}}
  \item{\code{rContrast}}{\code{sqrt(t^2/(t^2 + df))}; useful for power analyses}
}
}
\description{
This function allows to perform contrast analyses for independent samples
}
\examples{
set.seed(1)
nGroup <- 4
lambda <- matrix(c(-1,-1,1,
                   1,-1,1,-1,
                   1,1,-1,-1,1),
                 ncol = 4,
                 byrow=TRUE)
N <- 50
dat <- data.frame(
  x = rep(c(1:nGroup),each = N),
  y = c(rnorm(N,-1,1),rnorm(N),rnorm(N),rnorm(N,1,1))
)
contrast_independent(nGroup, lambda, dat)
# results are the same as the standard linear model or anova:

dat$c1 <- ifelse(dat$x==1 | dat$x==2, -1,1)
dat$c2 <- ifelse(dat$x==1 | dat$x==3, -1,1)
dat$c3 <- ifelse(dat$x==2 | dat$x==3, -1,1)
lmMod <- lm(y~c1+c2+c3,data=dat)
summary(lmMod)
anova(lmMod)

}
