% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastanalysis-package.R
\docType{package}
\name{contrastanalysis-package}
\alias{contrastanalysis-package}
\alias{contrastanalysis}
\title{contrastanalysis: Testing Single and Competing Contrasts}
\description{
This R-packages allows to perform contrast analyses for independent and dependent samples.
}
\author{
Mirka Henninger & Simone Malejka
}
