% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastCompare_dependent.R
\name{contrastCompare_dependent}
\alias{contrastCompare_dependent}
\title{Contrast analyses for dependent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
contrastCompare_dependent(nGroup, lambda1, lambda2, dat, testvalue = 0)
}
\arguments{
\item{nGroup}{Number of dependent / within-subject groups}

\item{lambda1}{a vector of contrast weights for Hypothesis 1}

\item{lambda2}{a vector of contrast weights for Hypothesis 2}

\item{dat}{a matrix or dataframe with nGroup columns; each row contains values for
one respondent;
each column contains values of the dependent variable in the respective within-
subject group}

\item{testvalue}{value under the Null hypothesis; if not specified, it is fixed to 0.}
}
\value{
a list with following entries:
\describe{
  \item{\code{Results}}{Results of the contrast analysis}
  \item{\code{Contrast Weights}}{Standardized contrast weights of the two original contrasts
  and difference between standardized contrast weights}
}
}
\description{
This function allows to directly compare two contrasts for dependent samples by
standardizing the contrast weights.
Please note that sample sizes must be equal between within-subject groups,
so no missings are allowed
}
\note{
A test favoring Hypothesis 1 is performed, hence a positive t-value indicate that the
contrast weights contained in lambda1 fit the data better than the contrast weights
contained in lambda2, and vice versa for a negative t-value.
}
\examples{
nGroup <- 4
N <- 50
lambda1 <- c(3,-1,-1,-1)
lambda2 <- c(0,2,-1,-1)
dat <- data.frame(
  group1 = sample(1:10, size = N, replace = TRUE),
  group2 = sample(1:15, size = N, replace = TRUE),
  group3 = sample(1:15, size = N, replace = TRUE),
  group4 = sample(1:20, size = N, replace = TRUE)
)
contrastCompare_dependent(nGroup, lambda1, lambda2, dat)

}
