% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastCompare_independent.R
\name{contrastCompare_independent}
\alias{contrastCompare_independent}
\title{Contrast analyses for independent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
contrastCompare_independent(nGroup, lambda1, lambda2, dat)
}
\arguments{
\item{nGroup}{Number of independent / between-subject groups}

\item{lambda1}{a vector of contrast weights for Hypothesis 1}

\item{lambda2}{a vector of contrast weights for Hypothesis 2}

\item{dat}{a matrix or dataframe with two columns; each row contains values for}

\item{dat}{a matrix or dataframe with two columns; each row contains values for
one respondents; the first column contains the group indicator, the second column
contains the dependent variable}
}
\value{
a dataframe with following entries for each of the contrasts:
\describe{
  \item{\code{SumsofSquares}}{Sums of Squares}
  \item{\code{F}}{F-values}
  \item{\code{estimate}}{Contrast estimates}
  \item{\code{t}}{t-values}
  \item{\code{p}}{two-tailed p-values}
  \item{\code{rEffectSize}}{Correlation between the dependent variable and the
  contrast weights}
  \item{\code{rAlerting}}{Correlation between group means and contrast weights}
  \item{\code{r2Aalerting}}{Squared \code{rAlerting}; can be interpreted similar to
  a determination coefficient as a measured of explained variance by the contrast
  \code{(SScontrast/SSbetween)}}
  \item{\code{rContrast}}{\code{sqrt(t^2/(t^2 + df))}; useful for power analyses}}
}
\description{
This function allows to directly compare two contrasts for independent samples by
standardizing the contrast weights
}
\note{
A test favoring Hypothesis 1 is performed, hence a positive t-value indicate that the
contrast weights contained in lambda1 fit the data better than the contrast weights
contained in lambda2, and vice versa for a negative t-value.
}
\examples{
# set.seed(1)
nGroup <- 4
lambda1 <- c(-1,-1,1,1)
lambda2 <- c(-3,-1,1,3)
dat <- data.frame(
  x = rep(c(1:4),each = 50),
  y = c(rnorm(50,-1,1),rnorm(50),rnorm(50),rnorm(50,1,1))
)
contrastCompare_independent(nGroup, lambda1, lambda2, dat)
}
