% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{working_memory}
\alias{working_memory}
\title{Working memory data}
\format{
a data frame with 6660 rows and 4 variables:
\describe{
  \item{ID}{the participant's id}
  \item{Trial}{trial indicator}
  \item{SetSize}{set size manipulation (within-subjects)}
  \item{confidence}{confidence rating}
}
}
\usage{
working_memory
}
\description{
This dataset originates from a study conducted at the University of Zurich
in the area of visual working memory. This is a vast simplification of
the original study for illustrative purposes.
The experiment consisted of 324 trials per participant. They were presented
with eather two, four, or eight colored squares (set size manipulation) and
had to reproduce the color shortly afterwards. Then they were asked to
judge their confidence on the given response.
}
\details{
For more details see: https://psyarxiv.com/xmf24/

For the entire dataset and analysis scripts see: https://osf.io/5mgzt/
}
\keyword{datasets}
