% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_independent.R
\name{compare_independent}
\alias{compare_independent}
\title{Contrast analyses for independent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
compare_independent(n_group, lambda1, lambda2, data)
}
\arguments{
\item{n_group}{Number of independent / between-subject groups}

\item{lambda1}{A vector of contrast weights for Hypothesis 1}

\item{lambda2}{A vector of contrast weights for Hypothesis 2}

\item{data}{A matrix or dataframe with two columns; each row contains values for
one respondents; the first column contains the group indicator, the second column
contains the dependent variable}
}
\value{
A list with following entries:
\describe{
  \item{\code{Results}}{Results of the contrast analysis}
  \item{\code{Contrast Weights}}{Standardized contrast weights of the two original
   contrasts and difference between standardized contrast weights}
}
}
\description{
This function allows to directly compare two contrasts for independent samples by
standardizing the contrast weights
}
\note{
A test favoring Hypothesis 1 is performed, hence a positive t-value indicate
 that the contrast weights contained in lambda1 fit the data better than the contrast
 weights contained in lambda2, and vice versa for a negative t-value.
}
\examples{
# load iris dataset
data("iris")
iris <- iris[,c("Species", "Petal.Length")]

# define lambda weights
lambda1 <- c(-1, 0, 1)  # H1: Iris versicolor is distinct to Iris setosa and Iris virginica
lambda2 <- c(-2, 1, 1)  # H2: Iris versicolor is more similar to Iris virginica

# perform contrast analysis
compare_independent(n_group=3, lambda1, lambda2, iris)

}
