% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_independent.R
\name{compare_independent}
\alias{compare_independent}
\title{Contrast analyses for independent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
compare_independent(nGroup, lambda1, lambda2, dat)
}
\arguments{
\item{nGroup}{Number of independent / between-subject groups}

\item{lambda1}{A vector of contrast weights for Hypothesis 1}

\item{lambda2}{A vector of contrast weights for Hypothesis 2}

\item{dat}{A matrix or dataframe with two columns; each row contains values for
one respondents; the first column contains the group indicator, the second column
contains the dependent variable}
}
\value{
A list with following entries:
\describe{
  \item{\code{Results}}{Results of the contrast analysis}
  \item{\code{Contrast Weights}}{Standardized contrast weights of the two original
   contrasts and difference between standardized contrast weights}
}
}
\description{
This function allows to directly compare two contrasts for independent samples by
standardizing the contrast weights
}
\note{
A test favoring Hypothesis 1 is performed, hence a positive t-value indicate
 that the contrast weights contained in lambda1 fit the data better than the contrast
 weights contained in lambda2, and vice versa for a negative t-value.
}
\examples{
set.seed(1)
nGroup <- 4
lambda1 <- c(-1,-1,1,1)
lambda2 <- c(-3,-1,1,3)
dat <- data.frame(
  x = rep(c(1:4),each = 50),
  y = c(rnorm(50,-1,1),rnorm(50),rnorm(50),rnorm(50,1,1))
)
compare_independent(nGroup, lambda1, lambda2, dat)

}
