% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_dependent.R
\name{contrast_dependent}
\alias{contrast_dependent}
\title{Contrast analyses for dependent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
contrast_dependent(nGroup, lambda, dat, testvalue = 0)
}
\arguments{
\item{nGroup}{Number of dependent / within-subject groups}

\item{lambda}{a matrix of contrast weights with contrasts in rows and groups in
columns}

\item{dat}{a matrix or dataframe with nGroup columns; each row contains values for
one respondent;
each column contains values of the dependent variable in the respective within-
subject group}

\item{testvalue}{value under the Null hypothesis; if not specified, it is fixed to 0.}
}
\value{
a dataframe with following entries for each of the contrasts:
\describe{
  \item{\code{SumsofSquares}}{Sums of Squares}
  \item{\code{F}}{F-values}
  \item{\code{estimate}}{Contrast estimates}
  \item{\code{t}}{t-values}
  \item{\code{p}}{two-tailed p-values}
  \item{\code{g}}{effect size g, a standardized distance measure}
}
}
\description{
This function allows to perform contrast analyses for dependent samples
}
\examples{
nGroup <- 4
N <- 50
lambda <- matrix(c(3,-1,-1,-1,
                   0,2,-1,-1,
                   0,0,-1,1),
                 ncol = nGroup,
                 byrow=TRUE)
dat <- data.frame(
  group1 = sample(1:10, size = N, replace = TRUE),
  group2 = sample(1:15, size = N, replace = TRUE),
  group3 = sample(1:15, size = N, replace = TRUE),
  group4 = sample(1:20, size = N, replace = TRUE)
)
contrast_dependent(nGroup, lambda,dat)

}
