% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_independent.R
\name{contrast_independent}
\alias{contrast_independent}
\title{Contrast analyses for independent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
contrast_independent(n_group, lambda, dat)
}
\arguments{
\item{n_group}{Number of independent / between-subject groups}

\item{lambda}{A matrix of contrast weights with contrasts in rows and
groups in columns}

\item{dat}{A matrix or dataframe with two columns; each row contains
values for one respondent;
the first column contains the group indicator, the second column contains
the dependent variable}
}
\value{
A dataframe with following entries for each of the contrasts:
\describe{
  \item{\code{SumsofSquares}}{Sums of Squares}
  \item{\code{df}}{Degrees of freedom for each contrast}
  \item{\code{F}}{F-values}
  \item{\code{contrast estimate}}{Contrast estimates}
  \item{\code{t}}{t-values}
  \item{\code{p}}{Two-tailed p-values}
  \item{\code{rEffectSize}}{Correlation between the dependent variable and
  the contrast weights}
  \item{\code{rAlerting}}{Correlation between group means and contrast weights}
  \item{\code{r2Aalerting}}{Squared \code{rAlerting}; can be interpreted
  similar to a determination coefficient as a measured of explained variance
  by the contrast
  \code{(SS_contrast/SSbetween)}}
  \item{\code{rContrast}}{\code{sqrt(t^2/(t^2 + df))}; useful for power analyses}
}
}
\description{
This function allows to perform contrast analyses for independent samples
}
\examples{
# load iris dataset
data("iris")
iris <- iris[,c("Species", "Petal.Length")]
n_group <- length(levels(iris$Species))

# define lambda weights
lambda <- matrix(c(
                -1,0,1, # H1: An increase from setosa over versicolor to virginica
                -2,1,1), # H2: Setosa has smaller Petal Length than versicolor and virginica
                ncol = n_group,
                byrow=TRUE)
# perform contrast analysis
contrast_independent(n_group, lambda, iris)


}
