% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_independent.R
\name{contrast_independent}
\alias{contrast_independent}
\title{Contrast analyses for independent samples}
\usage{
contrast_independent(x)
}
\description{
This function allows to perform contrast analyses for independent samples
}
