% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastanalysis-package.R
\docType{package}
\name{contrastanalysis-package}
\alias{contrastanalysis-package}
\alias{contrastanalysis}
\title{contrastanalysis: Testing Single and Competing Contrasts}
\description{
contrastanalysis: Testing Single and Competing Contrasts
}
\author{
Mirka Henninger & Simone Malejka
}
