% Generated by roxygen2: do not edit by hand
\name{contrastanalysis-package}
\alias{contrastanalysis-package}
\alias{contrastanalysis}
\docType{package}
\title{contrastanalysis: Contrast Analysis with Equivalence Testing}
\description{
Implements the Pillaud (2025) contrast analysis framework integrated
with Campbell (2023) TOST equivalence testing. The main function is
\code{\link{contrast_analysis}}.
}
