% Generated by roxygen2: do not edit by hand
\name{contrastanalysis-package}
\alias{contrastanalysis-package}
\alias{contrastanalysis}
\docType{package}
\title{contrastanalysis: Contrast Analysis with Equivalence Testing}
\description{
Implements contrast analysis integrated
with Campbell (2024) TOST equivalence testing. The main function is
\code{\link{contrast_analysis}}.
}
