% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastCompare_independent.R
\name{contrastCompare_independent}
\alias{contrastCompare_independent}
\title{Contrast analyses for independent samples}
\source{
Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)
}
\usage{
contrastCompare_independent(nGroup, lambda1, lambda2, dat)
}
\arguments{
\item{nGroup}{Number of independent / between-subject groups}

\item{lambda1}{a vector of contrast weights for Hypothesis 1}

\item{lambda2}{a vector of contrast weights for Hypothesis 2}

\item{dat}{a matrix or dataframe with two columns; each row contains values for
one respondents; the first column contains the group indicator, the second column
contains the dependent variable}
}
\value{
a dataframe with following entries: sums-of-squares, F-value, contrast estimate,
t value, ' two-tailed p value, and effect sizes r effect, r alerting, and r contrast;
A test favoring Hypothesis 1 is performed, hence a positive t-value indicate that the
contrast weights contained in lambda1 fit the data better than the contrast weights
contained in lambda2, and vice versa for a negative t-value.
}
\description{
This function allows to directly compare two contrasts for independent samples
}
\examples{
# set.seed(1)
nGroup <- 4
lambda1 <- c(-1,-1,1,1)
lambda2 <- c(-3,-1,1,3)
dat <- data.frame(
  x = rep(c(1:4),each = 50),
  y = c(rnorm(50,-1,1),rnorm(50),rnorm(50),rnorm(50,1,1))
)
contrastCompare_independent(nGroup, lambda1, lambda2, dat)
}
